Added threshold setting for image stack class
The threshold value allows you to threshold the minimum acceptable focus
measure value acceptable as a focus maximum. This is a crude implementation
of places pixels to the background if they don't have acceptable focus measure
functions. This needs to be improved if no discernable peak was found and or
if multiple peaks where found rather than a single peak.

Other changes were made and additions will be added to this commit message
later